passez une commande seulement si tu es connecter

+++ app/controllers/carts_controller.rb
@@ -43,12 +43,15 @@ class CartsController < ApplicationController
 
 
     @prix_total_float = "#{@prix_total_to_pay.to_i},#{@prix_total_to_pay.to_s.split(".")[1]}"
-    Order.create(user_id: current_user.id, prix_total: @prix_total_to_pay)
 
-    # vide le panier
+    if user_signed_in?
+      Order.create(user_id: current_user.id, prix_total: @prix_total_to_pay)
 
-    Store.where(cart_id: Cart.find_by(user_id: current_user.id)).each do |i|
-      i.destroy
+      # vide le panier
+
+      Store.where(cart_id: Cart.find_by(user_id: current_user.id)).each do |i|
+        i.destroy
+      end
     end
   end
